// web/src/api/models/appKeywords200.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * OpenApps API
 * Open source app intelligence toolkit API. Provides endpoints for app registration, DNA building, store listing tracking, and analytics.
 * OpenAPI spec version: 1.0.0
 */
import type { AppKeywords200Links } from './appKeywords200Links';
import type { AppKeywords200Meta } from './appKeywords200Meta';
import type { KeywordDensityResource } from './keywordDensityResource';

export type AppKeywords200 = {
  data?: KeywordDensityResource[];
  links?: AppKeywords200Links;
  meta?: AppKeywords200Meta;
};