// web/src/api/models/keywordDensityResource.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * OpenApps API
 * Open source app intelligence toolkit API. Provides endpoints for app registration, DNA building, store listing tracking, and analytics.
 * OpenAPI spec version: 1.0.0
 */

export interface KeywordDensityResource {
  locale?: string;
  ngram_size?: number;
  keyword: string;
  count: number;
  density: number;
}