// web/src/api/models/keywordCompareResourceAppsItem.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * OpenApps by MBZA API
 * Open source app intelligence toolkit API. Provides endpoints for app registration, DNA building, store listing tracking, and analytics.
 * OpenAPI spec version: 1.0.0
 */
import type { KeywordCompareResourceAppsItemVersionsItem } from './keywordCompareResourceAppsItemVersionsItem';

export type KeywordCompareResourceAppsItem = {
  id?: number;
  name?: string;
  /** @nullable */
  icon_url?: string | null;
  versions?: KeywordCompareResourceAppsItemVersionsItem[];
};