// web/src/api/models/keywordCompareResourceKeywords.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * OpenApps by MBZA API
 * Open source app intelligence toolkit API. Provides endpoints for app registration, DNA building, store listing tracking, and analytics.
 * OpenAPI spec version: 1.0.0
 */

export type KeywordCompareResourceKeywords = {[key: string]: {[key: string]: {
  count?: number;
  density?: number;
}}};