// web/src/api/models/appKeywordsOrder.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * OpenApps API
 * Open source app intelligence toolkit API. Provides endpoints for app registration, DNA building, store listing tracking, and analytics.
 * OpenAPI spec version: 1.0.0
 */

export type AppKeywordsOrder = typeof AppKeywordsOrder[keyof typeof AppKeywordsOrder];


export const AppKeywordsOrder = {
  asc: 'asc',
  desc: 'desc',
} as const;